requirements.txt
argcomplete==1.9.2
awscli==1.14.10
base58==0.2.4
boto3==1.4.8
botocore==1.8.14
certifi==2017.11.5
cfn-flip==0.3.0
chardet==3.0.4
click==6.7
colorama==0.3.7
Django==1.8
django-cors-headers==2.1.0
django-extensions==1.9.8
django-hstore==1.4.2
django-rest-framework==0.1.0
django-storages==1.6.5
djangorestframework==3.6.0
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.6
jmespath==0.9.3
jsonfield==2.0.2
kappa==0.6.0
lambda-packages==0.19.0
placebo==0.8.1
psycopg2==2.7.3.2
pyasn1==0.4.2
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2017.3
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
six==1.11.0
toml==0.9.3
tqdm==4.19.1
troposphere==2.1.2
typing==3.6.2
Unidecode==0.4.21
urllib3==1.22
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.45.1
